requirements.txt
fastapi
sqlalchemy
psycopg2-binary
openai
requests
uvicorn
python-dotenv

 